Added the ControllerSync Menu
Added all menus up to ControllerSync, just need to link the player class to the index of the respective controller so that when you play a mini game, the player can be linked to a controllable game object that will keep score for that player, once we figure that out, is a simple matter of starting the game, loading a random minigame and doing that until a game over. Note: add minigame scenes into an array of scenes. Everything in this commit has been tested by me and worked.